+++ components/secplus_gdo/__init__.py
@@ -26,8 +26,8 @@ from esphome.const import CONF_ID
 DEPENDENCIES = ["preferences"]
 MULTI_CONF = True
 
-konnectedgdo_ns = cg.esphome_ns.namespace("konnectedgdo")
-KONNECTEDGDO = konnectedgdo_ns.class_("GDOComponent", cg.Component)
+secplus_gdo_ns = cg.esphome_ns.namespace("secplus_gdo")
+SECPLUS_GDO = secplus_gdo_ns.class_("GDOComponent", cg.Component)
 
 CONF_OUTPUT_GDO = "output_gdo_pin"
 DEFAULT_OUTPUT_GDO = ("1")
@@ -36,11 +36,11 @@ DEFAULT_INPUT_GDO = ("2")
 CONF_INPUT_OBST = "input_obst_pin"
 DEFAULT_INPUT_OBST = ("-1")
 
-CONF_KONNECTEDGDO_ID = "konnectedgdo_id"
+CONF_SECPLUS_GDO_ID = "secplus_gdo_id"
 
 CONFIG_SCHEMA = cv.Schema(
     {
-        cv.GenerateID(): cv.declare_id(KONNECTEDGDO),
+        cv.GenerateID(): cv.declare_id(SECPLUS_GDO),
         cv.Required(CONF_OUTPUT_GDO): pins.gpio_output_pin_schema,
         cv.Required(CONF_INPUT_GDO): pins.gpio_input_pin_schema,
         cv.Optional(CONF_INPUT_OBST, default=DEFAULT_INPUT_OBST): cv.Any(
@@ -49,9 +49,9 @@ CONFIG_SCHEMA = cv.Schema(
     }
 ).extend(cv.COMPONENT_SCHEMA)
 
-KONNECTED_GDO_CONFIG_SCHEMA = cv.Schema(
+SECPLUS_GDO_CONFIG_SCHEMA = cv.Schema(
     {
-        cv.Required(CONF_KONNECTEDGDO_ID): cv.use_id(KONNECTEDGDO),
+        cv.Required(CONF_SECPLUS_GDO_ID): cv.use_id(SECPLUS_GDO),
     }
 )
