+++ components/secplus_gdo/gdolib.h
