+++ components/secplus_gdo/lock/__init__.py
@@ -22,17 +22,17 @@ import esphome.config_validation as cv
 from esphome.components import lock
 from esphome.const import CONF_ID
 
-from .. import KONNECTED_GDO_CONFIG_SCHEMA, konnectedgdo_ns, CONF_KONNECTEDGDO_ID
+from .. import SECPLUS_GDO_CONFIG_SCHEMA, secplus_gdo_ns, CONF_SECPLUS_GDO_ID
 
-DEPENDENCIES = ["konnectedgdo"]
+DEPENDENCIES = ["secplus_gdo"]
 
-GDOLock = konnectedgdo_ns.class_("GDOLock", lock.Lock, cg.Component)
+GDOLock = secplus_gdo_ns.class_("GDOLock", lock.Lock, cg.Component)
 
 CONFIG_SCHEMA = lock.LOCK_SCHEMA.extend(
     {
         cv.GenerateID(): cv.declare_id(GDOLock),
     }
-).extend(KONNECTED_GDO_CONFIG_SCHEMA)
+).extend(SECPLUS_GDO_CONFIG_SCHEMA)
 
 
 
@@ -40,6 +40,6 @@ async def to_code(config):
     var = cg.new_Pvariable(config[CONF_ID])
     await lock.register_lock(var, config)
     await cg.register_component(var, config)
-    parent = await cg.get_variable(config[CONF_KONNECTEDGDO_ID])
+    parent = await cg.get_variable(config[CONF_SECPLUS_GDO_ID])
     text = "std::bind(&" + str(GDOLock) + "::set_state," + str(config[CONF_ID]) + ",std::placeholders::_1)"
     cg.add(parent.register_lock(cg.RawExpression(text)))
