+++ components/secplus_gdo/sensor/gdo_sensor.h
@@ -18,12 +18,12 @@
 # pragma once
 
 #include "esphome/core/component.h"
-#include "esphome/components/binary_sensor/binary_sensor.h"
+#include "esphome/components/sensor/sensor.h"
 
 namespace esphome {
-namespace konnectedgdo {
+namespace secplus_gdo {
 
-class GDOBinarySensor : public binary_sensor::BinarySensor, public Component {};
+class GDOStat : public sensor::Sensor, public Component {};
 
-} // namespace konnectedgdo
+} // namespace secplus_gdo
 } // namespace esphome
