+++ components/secplus_gdo/switch/__init__.py
@@ -3,11 +3,11 @@ import esphome.config_validation as cv
 from esphome.components import switch
 from esphome.const import CONF_ID
 
-from .. import KONNECTED_GDO_CONFIG_SCHEMA, konnectedgdo_ns, CONF_KONNECTEDGDO_ID
+from .. import SECPLUS_GDO_CONFIG_SCHEMA, secplus_gdo_ns, CONF_SECPLUS_GDO_ID
 
-DEPENDENCIES = ["konnectedgdo"]
+DEPENDENCIES = ["secplus_gdo"]
 
-GDOSwitch = konnectedgdo_ns.class_("GDOSwitch", switch.Switch, cg.Component)
+GDOSwitch = secplus_gdo_ns.class_("GDOSwitch", switch.Switch, cg.Component)
 
 CONF_TYPE = "type"
 TYPES = {
@@ -22,7 +22,7 @@ CONFIG_SCHEMA = (
             cv.Required(CONF_TYPE): cv.enum(TYPES, lower=True),
         }
     )
-    .extend(KONNECTED_GDO_CONFIG_SCHEMA)
+    .extend(SECPLUS_GDO_CONFIG_SCHEMA)
 )
 
 
@@ -30,7 +30,7 @@ async def to_code(config):
     var = cg.new_Pvariable(config[CONF_ID])
     await switch.register_switch(var, config)
     await cg.register_component(var, config)
-    parent = await cg.get_variable(config[CONF_KONNECTEDGDO_ID])
+    parent = await cg.get_variable(config[CONF_SECPLUS_GDO_ID])
     fcall = str(parent) + "->" + str(TYPES[config[CONF_TYPE]])
     text = fcall + "(std::bind(&" + str(GDOSwitch) + "::write_state," + str(config[CONF_ID]) + ",std::placeholders::_1))"
     cg.add((cg.RawExpression(text)))
